Reorder the newrelic key addition and repo addition tasks
The Ansible apt_repository command updates the cache automatically
when adding a repo, unless you tell it otherwise.  But if you do that
before adding the repo key, to let apt know that the repo is trusted,
it is permanently untrustable.  And on Ubuntu 16.04, the SHA-1
algorithm was deprecated for repo keys, since it is too weak.

See also: https://discuss.newrelic.com/t/new-relic-gpg-key-is-no-longer-trusted-on-ubuntu-16-04-lts/36696/19